Relocate PrisonInteractionsHandler to engine package
